+++ src/DelayNodes.h
@@ -17,11 +17,11 @@ using namespace spotykach;
 class DelayNodes
 {
   public:
-    void init (float sampleRate, size_t blockSize, int numBands, int numProcs);
-  // Process a block. Inputs are per-band buffers (numBands_ x blockSize).
-  // Outputs are per-processor buffers (kMaxNumDelayProcs x blockSize).
-  // Only the first numProcs_ rows are written; the rest should be treated as zero.
-  void processBlockMono (float **inBand, float **treeOutputs, size_t ch, size_t blockSize);
+    void init (float sampleRate, size_t blockSize, size_t numBands, size_t numProcs);
+    // Process a block. Inputs are per-band buffers (numBands_ x blockSize).
+    // Outputs are per-processor buffers (kMaxNumDelayProcs x blockSize).
+    // Only the first numProcs_ rows are written; the rest should be treated as zero.
+    void processBlockMono (float **inBand, float **treeOutputs, size_t ch, size_t blockSize);
 
     int getNumBands () const { return numBands_; }
 
@@ -40,8 +40,8 @@ class DelayNodes
   private:
     float  sampleRate_ = 48000.0f;
     size_t blockSize_  = 16;
-    int    numBands_   = kMaxNutrientBands;
-    int    numProcs_   = kMaxNumDelayProcs;
+    size_t numBands_   = kMaxNutrientBands;
+    size_t numProcs_   = kMaxNumDelayProcs;
     float  stretch_    = 1.0f;
 
     // Tree density (0..1) maps linearly to number of active trees [1..numProcs_].
