+++ src/DiffusionControl.cpp
@@ -4,13 +4,13 @@
 constexpr float kPi = 3.14159265358979323846f;
 #endif
 
-void DiffusionControl::init (float sampleRate, int numBands)
+void DiffusionControl::init (float sampleRate, size_t numBands)
 {
   sampleRate_ = sampleRate;
-  numBands_   = std::max(1, std::min(numBands, kMaxNutrientBands));
-  for (int ch = 0; ch < spotykach::kNumberChannelsStereo; ++ch)
+  numBands_   = std::max(static_cast<size_t>(1), std::min(numBands, kMaxNutrientBands));
+  for (uint8_t ch = 0; ch < spotykach::kNumberChannelsStereo; ++ch)
   {
-    for (int band = 0; band < numBands_; ++band)
+    for (size_t band = 0; band < numBands_; ++band)
     {
       filters_[ch][band].Init(sampleRate_);
     }
@@ -20,9 +20,9 @@ void DiffusionControl::init (float sampleRate, int numBands)
 
 void DiffusionControl::setParameters (const Parameters &p)
 {
-  int   newBands = std::max(1, std::min(p.numActiveBands, kMaxNutrientBands));
-  float minFreq  = std::max(kMinFreq, p.centerFreq / 8);    // Three octaves below center frequency
-  float maxFreq  = std::min(kMaxFreq, p.centerFreq * 8);    // Three octaves above center frequency
+  size_t newBands = std::max(static_cast<size_t>(1), std::min(p.numActiveBands, kMaxNutrientBands));
+  float  minFreq  = std::max(kMinFreq, p.centerFreq / 8);    // Three octaves below center frequency
+  float  maxFreq  = std::min(kMaxFreq, p.centerFreq * 8);    // Three octaves above center frequency
   if ((newBands != numBands_) || (minFreq != kMinFreq) || (maxFreq != kMaxFreq))
   {
     numBands_ = newBands;
@@ -35,7 +35,7 @@ void DiffusionControl::setParameters (const Parameters &p)
 void DiffusionControl::getBandFrequencies (std::vector<float> &frequencies) const
 {
   frequencies.clear();
-  for (int band = 0; band < numBands_; ++band)
+  for (size_t band = 0; band < numBands_; ++band)
   {
     frequencies.push_back(centerFreqs_[band]);
   }
@@ -43,15 +43,15 @@ void DiffusionControl::getBandFrequencies (std::vector<float> &frequencies) cons
 
 void DiffusionControl::updateBandLayout ()
 {
-  for (int band = 0; band < numBands_; ++band)
+  for (size_t band = 0; band < numBands_; ++band)
   {
     float t            = (numBands_ == 1) ? 0.0f : static_cast<float>(band) / static_cast<float>(numBands_ - 1);
     centerFreqs_[band] = minFreq_ * powf(maxFreq_ / minFreq_, t);
   }
 
-  for (int ch = 0; ch < spotykach::kNumberChannelsStereo; ++ch)
+  for (uint8_t ch = 0; ch < spotykach::kNumberChannelsStereo; ++ch)
   {
-    for (int band = 0; band < numBands_; ++band)
+    for (size_t band = 0; band < numBands_; ++band)
     {
       filters_[ch][band].SetFreq(centerFreqs_[band]);
       filters_[ch][band].SetRes(0.707f);
@@ -61,7 +61,7 @@ void DiffusionControl::updateBandLayout ()
 
 void DiffusionControl::processBlockMono (const float *in, const uint8_t ch, float **outBand, size_t blockSize)
 {
-  for (int band = 0; band < numBands_; ++band)
+  for (size_t band = 0; band < numBands_; ++band)
   {
     float *o = outBand[band];
     auto  &f = filters_[ch][band];
